refactor: price를 Long으로, LocalDateTime을 Instant로, TIMESTAMP를 TIMESTAMPTZ로 전환

+++ productservice/src/main/java/com/trustamarket/productservice/domain/category/Category.java
@@ -55,7 +55,7 @@ public class Category {
                 ? this.inspectionThreshold
                 : CategoryThreshold.getThreshold(this.name);
     }
-    public boolean requiresInspection(int price) {
+    public boolean requiresInspection(Long price) {
         return switch (getEffectivePolicy()) {
             case ALWAYS      -> true;
             case NEVER       -> false;
